BC_Controller/requirements.txt
accelerate==0.21.0
beautifulsoup4==4.12.2
bitsandbytes==0.41.3
deepspeed==0.12.4
easydict==1.10
flax==0.7.5
ftfy==6.1.3
huggingface_hub==0.13.3
imageio==2.9.0
importlib_metadata==6.2.0
imwatermark==0.0.2
jax==0.4.21
k_diffusion==0.1.1.post1
librosa==0.10.1
matplotlib==3.7.3
minedojo==0.1
msgpack_python==0.5.6
note_seq==0.0.5
numpy==1.21.4
omegaconf==2.3.0
onnxruntime==1.16.3
opencv_python==4.8.0.74
opencv_python_headless==4.5.5.64
packaging==23.2
Pillow==9.5.0
Pillow==10.1.0
pytest==7.2.2
pytorch_lightning==2.0.9.post0
Requests==2.31.0
safetensors==0.3.1
scipy==1.9.1
skimage==0.0
tensorboardX==2.6.1
tensorboardX==2.6.2.2
torch==2.0.1
torchsde==0.2.6
torchvision==0.13.1
tqdm==4.64.1
tqdm==4.66.1
transformers==4.26.1
wandb==0.16.1
xformers==0.0.20
